+++ test/ignitor.spec.js
@@ -30,6 +30,8 @@ test.group('Ignitor', (group) => {
     fold.ioc.fake('Adonis/Src/Exception', () => {
       return { bind () {} }
     })
+
+    // process.removeAllListeners('unhandledRejection')
   })
 
   test('register app root', (assert) => {
@@ -95,12 +97,17 @@ test.group('Ignitor', (group) => {
     }
   })
 
-  test('register providers by requiring the app file', (assert) => {
-    const ignitor = new Ignitor(fold)
-    ignitor.appRoot(path.join(__dirname, './'))
-    ignitor._preLoadFiles = []
-    ignitor._startHttpServer = function () {}
-    ignitor.fireHttpServer()
+  test('register providers by requiring the app file', async (assert) => {
+    try {
+      const ignitor = new Ignitor(fold)
+      ignitor.appRoot(path.join(__dirname, './'))
+      ignitor._preLoadFiles = []
+      ignitor._startHttpServer = function () {}
+      ignitor._gracefullyShutDown = function () {}
+      await ignitor.fireHttpServer()
+    } catch (error) {
+      console.log(error)
+    }
   })
 
   test('emit before and after registered provider events', (assert) => {
@@ -115,6 +122,7 @@ test.group('Ignitor', (group) => {
     ignitor.appRoot(path.join(__dirname, './'))
     ignitor._preLoadFiles = []
     ignitor._startHttpServer = function () {}
+    ignitor._gracefullyShutDown = function () {}
     ignitor.fireHttpServer()
     assert.deepEqual(events, ['before', 'after'])
   })
@@ -143,6 +151,8 @@ test.group('Ignitor', (group) => {
     ignitor.appRoot(path.join(__dirname, './'))
     ignitor._preLoadFiles = []
     ignitor._startHttpServer = function () {}
+    ignitor._gracefullyShutDown = function () {}
+
     ignitor.fireHttpServer()
     assert.deepEqual(events, ['before', 'before 1', 'after', 'after 1'])
   })
@@ -152,6 +162,8 @@ test.group('Ignitor', (group) => {
     ignitor.appRoot(path.join(__dirname, './'))
     ignitor._preLoadFiles = []
     ignitor._startHttpServer = function () {}
+    ignitor._gracefullyShutDown = function () {}
+
     await ignitor.fireHttpServer()
     assert.deepEqual(fold.registrar._providers[0]._events, ['register', 'boot'])
   })
@@ -165,6 +177,8 @@ test.group('Ignitor', (group) => {
     ]
     ignitor._preLoadFiles = []
     ignitor._startHttpServer = function () {}
+    ignitor._gracefullyShutDown = function () {}
+
     await ignitor.fireHttpServer()
     assert.deepEqual(fold.registrar._providers[0]._events, ['register', 'boot'])
   })
@@ -185,6 +199,8 @@ test.group('Ignitor', (group) => {
     ignitor.appRoot(path.join(__dirname, './'))
     ignitor._preLoadFiles = []
     ignitor._startHttpServer = function () {}
+    ignitor._gracefullyShutDown = function () {}
+
     await ignitor.fireHttpServer()
     assert.deepEqual(events, ['before registered', 'after registered', 'before booted', 'after booted'])
   })
@@ -201,6 +217,8 @@ test.group('Ignitor', (group) => {
 
     ignitor._preLoadFiles = []
     ignitor._startHttpServer = function () {}
+    ignitor._gracefullyShutDown = function () {}
+
     await ignitor.fireHttpServer()
     assert.property(fold.ioc._aliases, 'Route')
     assert.property(fold.ioc._aliases, 'Server')
@@ -211,6 +229,8 @@ test.group('Ignitor', (group) => {
     ignitor.appRoot(path.join(__dirname, './'))
     ignitor._preLoadFiles = ['start/routes', 'start/events']
     ignitor._startHttpServer = function () {}
+    ignitor._gracefullyShutDown = function () {}
+
     await ignitor.fireHttpServer()
   })
 
@@ -225,6 +245,8 @@ test.group('Ignitor', (group) => {
     ignitor.appRoot(path.join(__dirname, './'))
     ignitor._preLoadFiles = []
     ignitor._startHttpServer = function () {}
+    ignitor._gracefullyShutDown = function () {}
+
     await ignitor.fireHttpServer()
     assert.deepEqual(events, ['before preloading', 'after preloading'])
   })
@@ -353,6 +375,11 @@ test.group('Ignitor', (group) => {
 
     class Server {
       listen (h, p, cb) { cb() }
+      getInstance () {
+        return {
+          once () {}
+        }
+      }
     }
 
     const events = []
@@ -370,5 +397,37 @@ test.group('Ignitor', (group) => {
     ignitor.appRoot(path.join(__dirname, './'))
     await ignitor.fireHttpServer()
     assert.deepEqual(events, ['before:httpServer', 'after:httpServer'])
+  })
+
+  test('bind custom http instance to adonis', async (assert) => {
+    const ignitor = new Ignitor(fold)
+    let customInstance = null
+
+    class Server {
+      setInstance (i) {
+        customInstance = i
+      }
+
+      getInstance () {
+        return customInstance
+      }
+
+      handle () {}
+
+      listen (h, p, cb) { cb() }
+    }
+
+    fold.ioc.fake('Adonis/Src/Server', () => new Server())
+    fold.ioc.fake('Adonis/Src/Env', () => new Env())
+
+    ignitor.appRoot(path.join(__dirname, './'))
+    let server = null
+
+    await ignitor.fireHttpServer(function (handler) {
+      server = require('http').createServer(handler)
+      return server
+    })
+
+    assert.deepEqual(customInstance, server)
   })
 })